Added denseset data-structure implementation.
The new denseset template container offers set semantics over an
open-addressed hashtable. Additionally cleaned up a few loose-ends
with the existing sparsemap/sparseset/densemap implementations and
unit tests.

This update also includes a fix for GCC not correctly emitting
the cmpunordps and cmpordps instructions when the corresponding
SSE intrinsics are used. The solution was to employ inline assembly
when GCC is the host compiler.