Added hashmap and hashset templated containers.
These new container templates are similar to densemap and denseset,
except that they assume the key and the hashcode are the same thing.
This is useful when data is pre-hashed. denseset is useful for
mapping handles/keys to array offsets that correspond to the offset
of the hashed key in the denseset.